+++ sdk/python/pulumi_azure/network/virtual_network_gateway_connection.py
@@ -59,7 +59,7 @@ class VirtualNetworkGatewayConnection(pulumi.CustomResource):
     The name of the resource group in which to
     create the connection Changing the name forces a new resource to be created.
     """
-    routing_weight: pulumi.Output[int]
+    routing_weight: pulumi.Output[float]
     """
     The routing weight. Defaults to `10`.
     """
@@ -122,7 +122,7 @@ class VirtualNetworkGatewayConnection(pulumi.CustomResource):
                in a different subscription.
         :param pulumi.Input[str] resource_group_name: The name of the resource group in which to
                create the connection Changing the name forces a new resource to be created.
-        :param pulumi.Input[int] routing_weight: The routing weight. Defaults to `10`.
+        :param pulumi.Input[float] routing_weight: The routing weight. Defaults to `10`.
         :param pulumi.Input[str] shared_key: The shared IPSec key. A key must be provided if a
                Site-to-Site or VNet-to-VNet connection is created whereas ExpressRoute
                connections do not need a shared key.
